Add responsive and move route
