+++ frontend/src/components/ProtectedRoute/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
-import { Navigate, Outlet, useLocation } from "react-router-dom";
+import { Navigate, Outlet } from "react-router-dom";
 import { useSelector } from "react-redux";
 import type { RootState } from "@/redux/store";
-import { isTokenExpired } from "./authHelper";
+import { isTokenExpired } from "../../utils/authHelper";
 
 const ProtectedRoute = () => {
   const {accessToken, exp} = useSelector((state: RootState) => state.auth);
