+++ frontend/src/pages/LoginPage/Components/LoginForm/component/SocialLoginButtons/SocialLoginButtons.tsx
@@ -1,10 +1,9 @@
-// components/SocialLoginButtons.tsx
 import { Stack, Button } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
 const GrayOutlinedButton = styled(Button)(({ theme }) => ({
-  color: theme.palette.common.black, // 字体颜色
-  borderColor: ' #E0E0E0', // 边框颜色
+  color: theme.palette.common.black,
+  borderColor: ' #E0E0E0', 
 }));
 
 const CustomGrayOutlinedButton = styled(GrayOutlinedButton)(({ theme }) => ({
